# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Chemidot B2B Chemical Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [buyer, supplier, admin]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, pending_verification]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersPage"
  /admin/users/{id}/status:
    put:
      operationId: adminUpdateUserStatus
      tags: [admin]
      summary: Update user status (ban/activate)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/suppliers/{id}/verify:
    put:
      operationId: adminVerifySupplier
      tags: [admin]
      summary: Verify/approve a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [buyer, supplier]
        firstName:
          type: string
        lastName:
          type: string
        companyName:
          type: string
        industry:
          type: string
        country:
          type: string
        phone:
          type: string
        commercialRegNumber:
          type: string
        warehouseLocation:
          type: string
      required: [email, password, role, firstName, lastName, companyName, country]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
          enum: [buyer, supplier, admin]
        firstName:
          type: string
        lastName:
          type: string
        companyName:
          type: string
        country:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, suspended, pending_verification]
        createdAt:
          type: string
      required: [id, email, role, firstName, lastName, companyName, country, status, createdAt]
...
    UsersPage:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [users, total, page, limit, totalPages]
    UpdateUserStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, suspended, pending_verification]
      required: [status]
...